[refactor] #53 - saved 페이지 코드 정리
